backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
flask-mysql-connector==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
Werkzeug==3.1.3
flask-sqlalchemy
